Rewrite core controller loop to use workqueue
This patch rewrites the controller loop to track outstanding tasks in
a workqueue.  This gives us (rate-limited) retries, and explicitly
deletes Secrets when we observe the corresponding SealedSecret going
away.

Note the latter isn't a proper garbage collection loop: We will leak
Secrets if the SealedSecret is deleted while the controller isn't
running.  This is a best-effort workaround until k8s supports
server-side garbage collection for TPR objects.