Added code for automatic key generation
Scheduler supports a periodic schedule with an early trigger
mechanism that is intended for use when a key is comprimised.

When a key name is generated, care is taken to ensure the keyName
is not already in use. This involves a call to the kubernetes
interface.

NOTE: Currently not integrated into the controller.