Added trigger go code and executable
This is a trigger activation implementation that is added to the
container for in container key-gen activation. This is intended to
replace the key-gen endpoint for access restriction to the trigger.

Change involves adding a second port opened by server.go that
should not be opened to the outside world. Also a new trigger file
is added to the makefile and copied into the controller image.