Removed keynames from user facing functionality
This commit addresses the concern that introducing keynames to
kubeseal and the sealedsecrets api introduces breaking changes and
exposes too many internal details to users.

In particular, removed all references and usages of keynames from
kubeseal and sealedsecret type defintions. Includes relevant
changes to controller to make it work with the changes.