Refactored key name and label naming scheme
This commit changes how the secrets storing public/private keys are
named and labelled. Names are still generated using the
generate-name option, however the prefix is now configurable by
a flag. The label is also configurable by flag separately. By
default the prefix is set to the label with a '-' added for
readability.

Defaults are set to 'sealed-secrets-key', same as current master
key name.