kubeseal: support loading cert from a well-known local file by default
This commit adds support to kubeseal to load the certificate / public
key from a well-known location by default:

	sealed-secrets.crt

We're proposing this change to open a conversation around supporting a
workflow that would be ideal for our developers. We'd like to document a
simple set of commands that each developer or team needs to run in their
local environment and/or CI pipeline to bootstrap the kubeseal command
for generating sealed secrets. In our case, we'd like to simply document
that developers need to run a command line this once in the workspace
where they will be running kubeseal:

```bash
kubeseal --controller-namespace=$FOO --fetch-cert > sealed-secrets.crt

kubeseal < secret.yaml > sealedsecret.yaml
```

The idea here is that if `--cert` is not specified when kubeseal runs,
and `sealed-secrets.crt` exists, it will use that file instead of
reaching out to the cluster - without developers having to specify
`--cert` every time. If we don't do this, then the error they see when
they forget `--cert`, and don't specify `--controller-namespace`, is a
little confusing to people who don't know how the sealed-secrets
controller works or where it runs:

	panic: Error fetching certificate: services "sealed-secrets-controller" not found

	goroutine 1 [running]:
	main.main()
		/home/travis/gopath/src/github.com/bitnami-labs/sealed-secrets/cmd/kubeseal/main.go:216 +0x3b9

Alternatively, for those of us who work on multiple clusters, we'd like
a slightly more advanced usage where an environment variable can be used to manage certificates for different clusters:

```bash
kubeseal --controller-namespace=$FOO --fetch-cert > ~/.sealed-secrets-prod.crt

SEALED_SECRETS_CERT=~/.sealed-secrets-prod.crt kubeseal < secret.yaml > sealedsecret.yaml
```

We're keen to discuss the workflow in more depth or to take another approach.
The idea is to leverage the existing `--fetch-cert` flag, not to break any
existing usage, and to slightly simplify the day-to-day workflow for developers
using kubeseal without knowing about the deployment or architecture of
`sealed-secrets`.