Added sort by key creation timestamp on startup
Controller now reads old keys on startup in order of their creation
times. This means the last key added is guaranteed to be the most
recent.

This commit also fixes an integration test issue when integration
tests are run against a controller with more than one key. Issue
was that fetching keys may not retrieve them in order, meaning the
tests were encrypting the test secret with a different key (via
kubeseal) that unencrypting (via retrieved private key), causing
some tests to fail. Sort in the tests fixes this issue.