Refactor of key registry to only curcial functions
This commit is essentially a cleanup, reducing the key registry
to its most basic requirements after refining other parts of key
rotation.

The registry now only keep a list of private keys and just the
latest cert. These are updates consistently throught the register
function. the latest private key is always at the end of the list
(appended whenever a new one is registered).