Refactor key book-keeping
Prepares the ground for addressing #185 and #199 without adding any feature.

The only small side effect is that now private keys are attempted in random order (Go map iteration order).
This could only affect users who have turned on the key rotation feature which, as mentioned in #137, is not yet ready for prime time in the first place.