Fix legacy key discovery
Legacy key discovery is wrong, see #265.

The TODO hints to the fact that the original plan was to label the legacy secret with `sealedsecrets.bitnami.com/sealed-secrets-key` before we create the subsequent ones, however that's brittle because there might be
cases where legacy and new secrets can come to coexist before such label has a chance to be applied (e.g. restores from a backup).

There is no reason why shouldn't just include the legacy secret if it exist.

Closes #265