fix(kubeseal): remove kubeconfig deps from recovery-unseal
`openCert()` in kubeseal triggers a kubeconfig check as it needs to fetch certificate information
from the running kubernetes cluster. Some people have used the `--cert` option as a way of
circumbent that check as it allows the kubeconfig check on `openCert()` to be bypassed by using a
previously backed up certificate. This commit applies a very basic "fix" by moving the `unseal`
command before the `openCert()` execution as it is a fairly self-sufficient command

fix #341