Refactor flag handling
The run method was too long and it was impossible to figure out what each parameter meant.
Furthermore a few flags were still accessed via global variables making it hard to reason about
and to test properly.

This is the first pass that just mechanically puts all the flag variables into a struct.