contrib/prometheus-mixin: Adjust alert to expose namespace and reason in labels
This changes how alert works to react on specific "reason" and in a specific namespace. It should allow easier debugging when alert fires. Additionally, I adjusted the alert name as it is not alerting on error rate (number of errors vs all requests) but a rate of change in metric value. Plus I adjusted tests as they were using unreal scenario.

Ideally, `reason` label should be included somehow in the alert description, but I didn't manage to do this in a way that would be grammatically correct :shrug:

Signed-off-by: paulfantom <pawel@krupa.net.pl>