Rebase of Refactor flag handling (#920)
The run method was too long and it was impossible to figure out what each parameter meant.
Furthermore a few flags were still accessed via global variables making it hard to reason about
and to test properly.

This refactor moves all flags into a Flags struct. It also moves some other global variables like the ClientConfig to a Config struct that also holds the Flags.

The result: no more global variables and run now just calls with the config and an output writer.