Fix Dockerfile to align with new GoReleaser Path format (#933)
GoReleaser includes in the path of the folder the version of amd64 (v1) and the arm (6) to generate the binaries. This new path format generates problems with the Dockerfile in the Release process. This solution allows us to create the container images in local and also in the Release process.