Make the kubeseal CLI have a skinny main (#939)
The CLI had way too much code within the main package.

This change moves most of the kubeseal CLI code to a package "pkg/kubeseal", leaving a simple "main" which is only concerned with CLI flags and other command line perks.

Note that the changes are minimal, and only functions affected by the split of code between main and the package are really affected. The rest is code copied to another package.

Other changes included:

- Flag fields needed to become public, and this are upper case now.
- There is a new NewConfig() function now.
- VERSION handling is handled by main completely and is out of the Flags struct now.

This change does not yet address shaping the kubeseal package code look like a proper library.