fix: add automount wrapper
Only turns off automount if we make a service account. The behavior in the provider I'm using also fails for default so maybe I'm doing it wrong. The existing flag being off would result in it being off for everyone. Do we need a more sophisticated way to enable/disable token creation?