Release the newest Sealed Secrets version using the Bitnami container release
With these changes, we are going to check if the Bitnami Container
Release process has generated the containers using the tag of with our
release. Then, we will generate the official release in our project.

Why this decision? Using the Bitnami Release process, the containers are
tested in several environment (OpenShift, AKS, etc) running our
integration tests.

From the user side, you will not see any change at all. We are
publishing the containers in the same registry and our chart and
controller will use same containers.

Signed-off-by: Alvaro Neira Ayuso <alvaron@vmware.com>