Release the newest Sealed Secrets version using the Bitnami container release (#1392)
With these changes, we are going to release the newest Sealed Secrets
version using the containers generated by the Bitnami Container Release
process using the tag of with our release. Then, we will generate the
official release in our project.

Why this decision? Using the Bitnami Release process, the containers are
tested in several environment (OpenShift, AKS, etc) running our
integration tests. Like this, the project will be covered in several
real environments that Sealed Secrets in running by the users.

From the user side, **you will not notice any change at all**. We are
publishing the containers in the same registry and our chart and
controller will use the same containers.

fixes #1393

---------

Signed-off-by: Alvaro Neira Ayuso <alvaron@vmware.com>