[helm] add rbac.proxier config (#1451)
**Description of the change**

Add options to the Helm chart to configure the `proxier` Role, and its
RoleBinding.

**Benefits**

This allows the user to override the default RoleBinding that grants
`system:authenticated` the `proxier` role. `system:authenticated` is not
a very safe option in a lot of clusters, so it's important to expose
this option.

**Possible drawbacks**

This was written to preserve existing behavior by default, so as to not
break compatibility. This isn't ideal as the existing behavior is
insecure on GKE (and possibly other contexts as well), but since it
doesn't directly expose anything damaging (but rather is a
defense-in-depth measure), it's probably OK to just highlight the
importance of changing this in the documentation.

**Applicable issues**
- partially addresses #1448

Signed-off-by: Alice Sawatzky <alice.sawatzky@farmersedge.ca>