feature: loadbalancerclass (#1545)
**Description of the change**

This allows to specify a loadbalancerclass on the created services.
From the kubernetes docs: 

> For a Service with type set to LoadBalancer, the
.spec.loadBalancerClass field enables you to use a load balancer
implementation other than the cloud provider default.
> 
> By default, .spec.loadBalancerClass is not set and a LoadBalancer type
of Service uses the cloud provider's default load balancer
implementation if the cluster is configured with a cloud provider using
the --cloud-provider component flag.
> 
> If you specify .spec.loadBalancerClass, it is assumed that a load
balancer implementation that matches the specified class is watching for
Services. Any default load balancer implementation (for example, the one
provided by the cloud provider) will ignore Services that have this
field set. spec.loadBalancerClass can be set on a Service of type
LoadBalancer only. Once set, it cannot be changed. The value of
spec.loadBalancerClass must be a label-style identifier, with an
optional prefix such as "internal-vip" or "example.com/internal-vip".
Unprefixed names are reserved for end-users.

---------

Signed-off-by: Matthias Teich <matthiasteich83@gmail.com>
Signed-off-by: Alvaro Neira Ayuso <alvaroneay@gmail.com>