Fix register a key using secret creationTimestamp instead of certificate validity timestamp (#1681)
<!--
Before you open the request please review the following guidelines and
tips to help it be more easily integrated:

 - Describe the scope of your change - i.e. what the change does.
 - Describe any known limitations with your change.
- Please run any tests or examples that can exercise your modified code.

Thank you for contributing! We will try to test and integrate the change
as soon as we can, but be aware we have many GitHub repositories to
manage and can't immediately respond to every request. There is no need
to bump or check in on a pull request (it will clutter the discussion of
the request).

Also don't be worried if the request is closed or not integrated
sometimes the priorities of Bitnami might not match the priorities of
the pull request. Don't fret, the open source community thrives on forks
and GitHub makes it easy to keep your changes in a forked repo.
 -->

**Description of the change**

Register a key and order it based on its secret creation timestamp
instead of its cert starting validity (NotBefore attribute)

<!-- Describe the scope of your change - i.e. what the change does. -->

**Benefits**

<!-- What benefits will be realized by the code change? -->

In case we are bringing our own certificate, if the certificate is
created before sealed secret installation but applied after the
installation, still the latest certificate used will be the certificate
installed by sealed secret during its initialization instead of the
latest secret certificate created.

**Possible drawbacks**

<!-- Describe any known limitations with your change -->

**Applicable issues**

<!-- Enter any applicable Issues here (You can reference an issue using
#) -->
- fixes # https://github.com/bitnami-labs/sealed-secrets/issues/1639

**Additional information**

<!-- If there's anything else that's important and relevant to your pull
request, mention that information here.-->

---------

Signed-off-by: Jérôme GARCIA <35989114+sybernatus@users.noreply.github.com>
Signed-off-by: Sybernatus <jerome.je.garcia@gmail.com>
Signed-off-by: Jérôme GARCIA <jerome.jerom.garcia.external@airbus.com>