helm: add watch for secrets (#1758)
<!--
Before you open the request please review the following guidelines and
tips to help it be more easily integrated:

 - Describe the scope of your change - i.e. what the change does.
 - Describe any known limitations with your change.
- Please run any tests or examples that can exercise your modified code.

Thank you for contributing! We will try to test and integrate the change
as soon as we can, but be aware we have many GitHub repositories to
manage and can't immediately respond to every request. There is no need
to bump or check in on a pull request (it will clutter the discussion of
the request).

Also don't be worried if the request is closed or not integrated
sometimes the priorities of Bitnami might not match the priorities of
the pull request. Don't fret, the open source community thrives on forks
and GitHub makes it easy to keep your changes in a forked repo.
 -->

**Description of the change**

<!-- Describe the scope of your change - i.e. what the change does. -->

MR [#1737](https://github.com/bitnami-labs/sealed-secrets/pull/1737)
added support for `--watch-for-secrets` but did not add it to the helm
chart.

**Benefits**

<!-- What benefits will be realized by the code change? -->

Will be able to enable --watch-for-secrets with helm without using args.

**Possible drawbacks**

<!-- Describe any known limitations with your change -->
--watch-for-secrets is still beta.

**Additional information**

<!-- If there's anything else that's important and relevant to your pull
request, mention that information here.-->

Signed-off-by: John Allberg <john@ayoy.se>