Compile again binaries and create containers internally
As the community knows, Bitnami will start to reduce the container
generation for several products. Sealed Secrets is a independent project
and for this, we need to start generating again our containers.

We will continue pushing them to the same registry. For this reason, ***the
users should not be affected at all***.

Signed-off-by: Alvaro Neira Ayuso <alvaro.neira@broadcom.com>